requirements.txt
Django==1.11
django-localflavor==1.4.1
psycopg2==2.7.1
